tried to get a test coded
