+++ lib/shapes.js
@@ -26,26 +26,4 @@ class Square extends Shape {
     }
 };
 
-function generateShapes({ shape, shapeColor, text, textColor }) {
-    let shapeCreate;
-
-    if (shape === 'circle') {
-        shapeCreate = new Circle({ shapeColor, text, textColor });
-    }
-    if (shape === 'triangle') {
-        shapeCreate = new Triangle({ shapeColor, text, textColor });
-    }
-    if (shape === 'square') {
-        shapeCreate = new Square({ shapeColor, text, textColor });
-    }
-
-    return `<svg version="1.1"
-            width="300" height="200"
-            xmlns="http://www.w3.org/2000/svg">
-
-        ${shapeCreate.render()}
-
-    </svg>`;
-};
-
-module.exports = generateShapes;
+module.exports = { Circle, Triangle, Square };
