+++ lib/shapes.test.js
@@ -1,0 +1,3 @@
+const { Circle, Triangle, Square } = require('./shapes');
+
+describe('Circle,
