+++ package.json
@@ -15,5 +15,8 @@
   "license": "ISC",
   "dependencies": {
     "inquirer": "^8.2.4"
+  },
+  "devDependencies": {
+    "jest": "^29.6.3"
   }
 }
